% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_priors.R
\name{compute_priors}
\alias{compute_priors}
\title{Get prior means and variances using empirical bayes approach}
\usage{
compute_priors(df, bigSNP_file)
}
\arguments{
\item{inFile}{Path to the input file}
}
\value{
A dataframe of summary statistics with columns chr, pos, a0, a1, beta
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
}
