% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights.R
\name{create_weights}
\alias{create_weights}
\title{Create weights for LDSC}
\usage{
create_weights(df)
}
\arguments{
\item{inFile}{Path to the input file}
}
\value{
A dataframe of summary statistics with columns chr, pos, a0, a1, beta
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
}
